Add a settings page to notes to handle wider permission piece
