+++ apps/game-master-client/app/components/entity-toolbar.tsx
@@ -1,24 +1,27 @@
-import { Share1Icon, TrashIcon } from "@radix-ui/react-icons";
-import { Form, useSubmit } from "@remix-run/react";
+import { Share1Icon } from "@radix-ui/react-icons";
+import { Form, useNavigate, useSubmit } from "@remix-run/react";
 import { Button } from "~/components/ui/button";
 import { Popover, PopoverDialog, PopoverTrigger } from "~/components/ui/popover";
 import { JollyTextField } from "~/components/ui/textfield";
 import { Toolbar } from "~/components/ui/toolbar";
+import { JollyMenu, MenuItem } from "./ui/menu";
+import { useState } from "react";
 
 export function EntityToolbar() {
 	const submit = useSubmit();
+	const navigate = useNavigate();
 	return (
-		<Toolbar>
-			<Button
-				variant={"destructive"}
-				size={"icon"}
-				onPress={() => submit({}, { method: "delete" })}
-				aria-label="Delete note"
-			>
-				<TrashIcon />
-			</Button>
-			<DuplicateEntityPopover />
-		</Toolbar>
+		<>
+			<Toolbar>
+				<JollyMenu label="Menu" variant="outline">
+					<MenuItem onAction={() => submit({}, { method: "delete" })}>Delete</MenuItem>
+					<MenuItem onAction={() => navigate("settings", { relative: "path" })}>
+						Settings..
+					</MenuItem>
+				</JollyMenu>
+				<DuplicateEntityPopover />
+			</Toolbar>
+		</>
 	);
 }
