+++ apps/game-master-client/app/routes/_app.games.$gameId.notes.$noteId._index/components/note-sidebar.tsx
