+++ apps/game-master-client/app/routes/_app.games.$gameId.notes.$noteId.settings/route.tsx
@@ -1,0 +1,5 @@
+export default function NoteSettingsRoute() {
+  return (
+    <div>This is the note settings page</div>
+  )
+}
